+++ .ebextensions/python.config
