--- DL_Models/LSTM/lstm_utils.py
@@ -1,20 +1,0 @@
-import numpy as np
-
-
-def min_max_scaler(data):
-    min_val = np.min(data, axis=0)
-    max_val = np.max(data, axis=0)
-    return (data - min_val) / (max_val - min_val), min_val, max_val
-
-
-def inverse_min_max_scaler(scaled_data, min_val, max_val):
-    return scaled_data * (max_val - min_val) + min_val
-
-
-def create_dataset(dataset, look_back):
-    dataX, dataY = [], []
-    for i in range(len(dataset) - look_back):
-        a = dataset[i:(i + look_back), :]
-        dataX.append(a)
-        dataY.append(dataset[i + look_back, 0])
-    return np.array(dataX), np.array(dataY)
