--- buildspec.yml
@@ -1,31 +1,0 @@
-version: 0.2
-
-phases:
-  install:
-    runtime-versions:
-      python: 3.10  # Update to match your project's Python version
-    commands:
-      - echo Installing Python dependencies...
-      - pip install -r requirements.txt  # Install dependencies from requirements.txt
-
-  pre_build:
-    commands:
-      - echo Running pre-build steps...
-      # Add any pre-build commands here (like database migrations or setup scripts)
-
-  build:
-    commands:
-      - echo Running tests...
-
-  post_build:
-    commands:
-      - echo Build completed on `date`
-      # Any post-build steps or cleanup
-
-artifacts:
-  files:
-    - '**/*'  # Specify files and directories to package as build output
-
-cache:
-  paths:
-    - '/root/.cache/pip'  # Cache pip packages to speed up future builds
