+++ requirements.txt
@@ -1,13 +1,7 @@
+boto3==1.28.85
+Flask==2.2.3
+Flask_Cors==4.0.0
 matplotlib==3.7.4
-numpy==1.24.3
+Models==0.9.7
 pandas==1.5.3
-scikit_learn==1.3.0
-seaborn==0.13.0
-statsmodels==0.14.0
-torch==2.0.1
-xgboost==2.0.0
-scikit-learn~=1.3.0
-boto3~=1.28.66
-python-dotenv~=1.0.0
-flask~=3.0.0
-requests~=2.30.0
+Requests==2.31.0
